// src/types/HealthDataTypes.ts
export enum HealthDataType {
    WalkingAsymmetryPercentage = 'HKQuantityTypeIdentifierWalkingAsymmetryPercentage',
    FlightsClimbed = 'HKQuantityTypeIdentifierFlightsClimbed',
    HeadphoneAudioExposure = 'HKQuantityTypeIdentifierHeadphoneAudioExposure',
    AppleWalkingSteadiness = 'HKQuantityTypeIdentifierAppleWalkingSteadiness',
    WalkingSpeed = 'HKQuantityTypeIdentifierWalkingSpeed',
    BodyMass = 'HKQuantityTypeIdentifierBodyMass',
    Height = 'HKQuantityTypeIdentifierHeight',
    ActiveEnergyBurned = 'HKQuantityTypeIdentifierActiveEnergyBurned',
    DistanceWalkingRunning = 'HKQuantityTypeIdentifierDistanceWalkingRunning',
    SleepDurationGoal = 'HKDataTypeSleepDurationGoal',
    SleepAnalysis = 'HKCategoryTypeIdentifierSleepAnalysis',
    BasalEnergyBurned = 'HKQuantityTypeIdentifierBasalEnergyBurned',
    StepCount = 'HKQuantityTypeIdentifierStepCount',
    WalkingDoubleSupportPercentage = 'HKQuantityTypeIdentifierWalkingDoubleSupportPercentage',
    WalkingStepLength = 'HKQuantityTypeIdentifierWalkingStepLength',
    HeartRateVariabilitySDNN = 'HKQuantityTypeIdentifierHeartRateVariabilitySDNN',
    PhysicalEffort = 'HKQuantityTypeIdentifierPhysicalEffort',
    HeartRate = 'HKQuantityTypeIdentifierHeartRate',
    RespiratoryRate = 'HKQuantityTypeIdentifierRespiratoryRate',
    MindfulSession = 'HKCategoryTypeIdentifierMindfulSession',
    CervicalMucusQuality = 'HKCategoryTypeIdentifierCervicalMucusQuality',
    OxygenSaturation = 'HKQuantityTypeIdentifierOxygenSaturation',
}